Show AllGames title shadow only when visible
