+++ client/package.json
@@ -33,6 +33,7 @@
     "@fortawesome/free-regular-svg-icons": "5.15.3",
     "@fortawesome/free-solid-svg-icons": "5.15.3",
     "@fortawesome/react-fontawesome": "0.1.14",
+    "@react-hook/intersection-observer": "3.1.1",
     "axios": "0.21.1",
     "babel-plugin-styled-components": "1.13.2",
     "browserslist": "4.16.6",
